+++ apps/demos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # xnet-demos
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [[`c7ef045`](https://github.com/crs48/xNet/commit/c7ef0456bfc75b5813d8a9d34f465f13a1e088ae)]:
+  - @xnetjs/react@2.5.0
+  - @xnetjs/data@2.5.0
+  - @xnetjs/devtools@0.1.0
+  - @xnetjs/identity@2.5.0
+
 ## 0.1.6
 
 ### Patch Changes